Replaced all double floating point precision maths with single floating point as we don't need the precision and it's faster. Added diagonal moves to the graph.

+++ C++ Library/PathfindingLib/Examples/Target.cpp
@@ -2,7 +2,7 @@
 #include "Target.h"
 
 
-Target::Target(double x, double y)
+Target::Target(float x, float y)
 {
 	this->x = x;
 	this->y = y;
