+++ C++ Library/PathfindingLib/PathfindingLib/PSystem.h
@@ -15,15 +15,15 @@ public:
 	PSystem(PSystem const&) = delete;
 	void operator=(PSystem const&) = delete;
 	// Graph Management
-	void InitGraph(unsigned int layer_id, blaze::StaticVector<double, 3> origin, blaze::StaticVector<double, 3> dimensions, double scale);
+	void InitGraph(unsigned int layer_id, blaze::StaticVector<float, 3> origin, blaze::StaticVector<float, 3> dimensions, float scale);
 	// Unit Management
-	unsigned int CreateUnit(blaze::StaticVector<double, 3> position, unsigned int layer_id);
-	void UpdateUnitTarget(unsigned int id, blaze::StaticVector<double, 3> target);
-	void UpdateUnitPosition(unsigned int id, blaze::StaticVector<double, 3> position);
-	void UpdateUnitHeading(unsigned int id, blaze::StaticVector<double, 3> heading);
+	unsigned int CreateUnit(blaze::StaticVector<float, 3> position, unsigned int layer_id);
+	void UpdateUnitTarget(unsigned int id, blaze::StaticVector<float, 3> target);
+	void UpdateUnitPosition(unsigned int id, blaze::StaticVector<float, 3> position);
+	void UpdateUnitHeading(unsigned int id, blaze::StaticVector<float, 3> heading);
 	void DestroyUnit(unsigned int id);
 	// Unit Interaction
-	blaze::StaticVector<double, 3> GetUnitForce(unsigned int id);
+	blaze::StaticVector<float, 3> GetUnitForce(unsigned int id);
 	// Layer Management
 	void CreateLayer(unsigned int layer_id);
 private:
@@ -34,11 +34,11 @@ private:
 	std::vector<unsigned int> free_unit_ids;
 	PUnit* GetUnit(unsigned int unit_id);
 	// Unit Interaction
-	double cohesion_factor = 0.001;
-	double separation_factor = 1;
-	double alignment_factor = 0 /*0.1*/;
-	double following_factor = 0.01;
-	double target_factor = 0.01;
+	float cohesion_factor = 0.001f;
+	float separation_factor = 1;
+	float alignment_factor = 0 /*0.1*/;
+	float following_factor = 0.01f;
+	float target_factor = 0.01f;
 	// Layer Management
 	std::unordered_map<unsigned int, PUnitLayer*> layers;
 	std::unordered_map<unsigned int, unsigned int> layer_allocation;
